Move inmemory_backend
