+++ ydb/library/actors/core/subsystems/inmemory_metrics.h
@@ -1,15 +1,9 @@
 #pragma once
 
 #include <ydb/library/actors/core/subsystem.h>
-#include <ydb/library/actors/metrics/line_base.h>
-#include <ydb/library/actors/metrics/line.h>
-#include <ydb/library/actors/metrics/line_storage.h>
-#include <ydb/library/actors/metrics/lines/raw_line_frontend.h>
-#include <ydb/library/actors/metrics/lines/on_change_line_frontend.h>
-#include <util/generic/function.h>
+#include <ydb/library/actors/metrics/inmemory_backend.h>
 
 #include <memory>
-#include <utility>
 
 namespace NActors {
     class TActorSystem;
@@ -19,7 +13,7 @@ namespace NActors {
     std::unique_ptr<TInMemoryMetricsRegistry> MakeInMemoryMetricsRegistry(TInMemoryMetricsConfig config);
     IActor* CreateInMemoryMetricsStatsActor(TDuration interval = TDuration::Seconds(1));
 
-    class TInMemoryMetricsRegistry : public ISubSystem, public TLineWriteBackend {
+    class TInMemoryMetricsRegistry : public ISubSystem {
     public:
         explicit TInMemoryMetricsRegistry(TInMemoryMetricsConfig config);
         ~TInMemoryMetricsRegistry() override;
@@ -30,7 +24,7 @@ namespace NActors {
         TLine<TRawLineFrontend<>> CreateLine(TStringBuf name, std::span<const TLabel> labels);
         template<class TFrontend>
         TLine<TFrontend> CreateLine(TStringBuf name, std::span<const TLabel> labels, const typename TFrontend::TConfig& config = {}) {
-            return TLine<TFrontend>(this, CreateLineWithMeta(name, labels, TFrontend::MakeMeta(config)));
+            return Backend.CreateLine<TFrontend>(name, labels, config);
         }
         void SetCommonLabels(std::span<const TLabel> labels);
         TVector<TLabel> GetCommonLabels() const;
@@ -42,33 +36,11 @@ namespace NActors {
         const TInMemoryMetricsConfig& GetConfig() const noexcept;
 
     private:
-        template<class TFrontend>
-        friend class TLine;
-        friend struct TSnapshotData;
-
-        class TImpl;
-
         void OnAfterStart(TActorSystem&) override;
         void OnBeforeStop(TActorSystem&) override;
-        TLineWriterState* CreateLineWithMeta(TStringBuf name, std::span<const TLabel> labels, const TLineMeta& meta);
-        void CloseLine(TLineWriterState* writer) noexcept override;
-        bool AppendStoredRecord(TLineWriterState* writer, const TStoredRecord& record) noexcept override;
-        NHPTimer::STime CurrentTimestampTs() const noexcept override;
-        TLinePublishState GetPublishState(const TLineWriterState* writer) const noexcept override;
-        ui32 GetLineId(const TLineWriterState* writer) const noexcept override;
-        void MarkObserved(TLineWriterState* writer, NHPTimer::STime nowTs) noexcept override;
-        void MarkPublished(TLineWriterState* writer, ui64 value, NHPTimer::STime nowTs) noexcept override;
-        TChunk* TryAcquireFreeChunk();
-        TChunk* TryStealOldestChunk();
-        void PublishSealedChunk(TChunk* chunk);
-        void RetireChunk(TChunk* chunk);
-        void ReturnChunkToFree(TChunk* chunk);
-        void ReleasePinnedChunk(TChunk* chunk) noexcept override;
-        void MaybeDropClosedLine(TLineReader* line);
-        bool IsMetricAllowed(TStringBuf name) const noexcept;
 
     private:
-        std::unique_ptr<TImpl> Impl;
+        TInMemoryMetricsBackend Backend;
     };
 
     TInMemoryMetricsRegistry* GetInMemoryMetrics(TActorSystem& actorSystem);
